@@ -1,19 +1,19 @@
 name: simple_markdown_editor
 description: A TextField Widget that allow you to convert easily what's in the
   TextField to Markdown.
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 0.2.0
 homepage: https://github.com/zahniar88/simple_markdown_editor
 repository: https://github.com/zahniar88/simple_markdown_editor
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: '>=3.2.0 <4.0.0'
 
 dependencies:
   expandable: ^5.0.1
   flutter:
     sdk: flutter
-  flutter_markdown: ^0.6.6
+  flutter_markdown: ^0.6.22+1
   font_awesome_flutter: ^10.7.0
   markdown: ^7.2.2
 
@@ -25,7 +25,11 @@ dev_dependencies:
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 # The following section is specific to Flutter.
-flutter: null
+flutter:
+  # The following line ensures that the Material Icons font is
+  # included with your application, so that you can use the icons in
+  # the material Icons class.
+  uses-material-design: true
 
 # To add assets to your package, add an assets section, like this:
 # assets:
